+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+0.9.0
+----
+Author: Tony DiCola
+- Added REST API support for all feed, data, group create/get/update/delete APIs.
+- Added explicit data model classes.
+- Added many integration tests to verify client & service.
+- Added docstrings to all public functions and classes.
+- Ported to work with Python 2 and 3.
+
 0.0.1
 ----
 Initial Changelog